clients/go/requirements-test.txt
chromadb>=1.0.0,<2.0.0
numpy<2.0
onnxruntime>=1.15.0,<2.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0
